+++ lib/aws-record-generator.rb
@@ -15,6 +15,7 @@ require_relative 'generators/generated_attribute'
 require_relative 'generators/secondary_index'
 require_relative 'generators/test_helper'
 require_relative 'generators/aws_record/model/model_generator'
+require_relative 'generators/aws_record/base'
 
 require 'aws-record'
 require 'rails/railtie'
