+++ lib/generators/aws_record/scaffold_controller/templates/api_controller.rb.tt
@@ -46,7 +46,7 @@ class <%= controller_class_name %>Controller < ApplicationController
   private
     # Use callbacks to share common setup or constraints between actions.
     def set_<%= singular_table_name %>
-      @<%= singular_table_name %> = <%= orm_class.find(class_name, "params[:id]") %>
+      @<%= singular_table_name %> = <%= orm_class.find(class_name, attributes) %>
     end
 
     # Only allow a trusted parameter "white list" through.
