+++ lib/generators/generated_attribute.rb
@@ -20,6 +20,32 @@ module AwsRecord
       attr_reader :name, :type
       attr_accessor :options
 
+      def column_name
+        @name
+      end
+
+      def password_digest?
+        false
+      end
+
+      def polymorphic?
+        false
+      end
+
+      def human_name
+        name.humanize
+      end
+
+      def field_type
+        case @type
+          when :integer_attr then :number_field
+          when :date_attr then :date_select
+          when :datetime_attr then :datetime_select
+          when :boolean_attr then :check_box
+          else :text_field
+        end
+      end
+
       class << self
 
         def parse(field_definition)
